requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aio-pika==8.2.5
aiormq==6.4.2; python_version >= '3.7'
aioschedule==0.5.2
alembic==1.8.1
anyio==3.6.2; python_full_version >= '3.6.2'
beautifulsoup4==4.11.1; python_version >= '3.6'
bs4==0.0.1
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.9.24; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
dnspython==2.2.1; python_version >= '3.6' and python_version < '4.0'
email-validator==1.3.0
fastapi-utils==0.2.1
fastapi[all]==0.87.0
google-api-core==2.10.2; python_version >= '3.7'
google-api-python-client==2.66.0
google-auth-httplib2==0.1.0
google-auth==2.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.57.0; python_version >= '3.7'
gunicorn==20.1.0
h11==0.14.0; python_version >= '3.7'
httpcore==0.16.1; python_version >= '3.7'
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.1
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
multidict==6.0.2; python_version >= '3.7'
numpy==1.23.5; python_version >= '3.10'
oauth2client==4.1.3
orjson==3.8.2
pamqp==3.2.1; python_version >= '3.7'
pandas==1.5.1
pika==1.3.1
protobuf==4.21.9; python_version >= '3.7'
psycopg2==2.9.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.10.2; python_version >= '3.7'
pydrive==1.3.1
pyparsing==3.0.9; python_version > '3.0'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
pyyaml==6.0
requests==2.28.1; python_version >= '3.7' and python_version < '4.0'
rfc3986[idna2008]==1.5.0
rsa==4.9; python_version >= '3.6' and python_version < '4.0'
schedule==1.1.0
setuptools==65.6.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlalchemy==1.4.44
starlette==0.21.0; python_version >= '3.7'
typing-extensions==4.4.0; python_version >= '3.7'
ujson==5.5.0
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4.0'
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.1; python_version >= '3.7'
